+++ src/flyingcircus/service/appmesh.py
@@ -1,0 +1,12 @@
+"""General-use classes to interact with the AppMesh service through CloudFormation.
+
+See Also:
+    `AWS developer guide for AppMesh
+    <https://docs.aws.amazon.com/app-mesh/?id=docs_gateway>`_
+"""
+
+# noinspection PyUnresolvedReferences
+from .._raw import appmesh as _raw
+
+# noinspection PyUnresolvedReferences
+from .._raw.appmesh import *